+++ gears/gears.cpp
@@ -83,7 +83,7 @@ public:
 		renderPassBeginInfo.clearValueCount = 2;
 		renderPassBeginInfo.pClearValues = clearValues;
 
-		for (int32_t i = 0; i < drawCmdBuffers.size(); ++i)
+		for (size_t i = 0; i < drawCmdBuffers.size(); ++i)
 		{
 			renderPassBeginInfo.framebuffer = frameBuffers[i];
 
@@ -132,7 +132,7 @@ public:
 		std::vector<float> rotationOffsets = { 0.0f, -9.0f, -30.0f };
 
 		gears.resize(positions.size());
-		for (int32_t i = 0; i < gears.size(); ++i)
+		for (size_t i = 0; i < gears.size(); ++i)
 		{
 			GearInfo gearInfo = {};
 			gearInfo.innerRadius = innerRadiuses[i];
